% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_cache.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear the local HMRC data cache}
\usage{
clear_cache(max_age_days = NULL)
}
\arguments{
\item{max_age_days}{Numeric or \code{NULL}. If \code{NULL} (default), all cached files
are removed. If a number, only files last modified more than that many
days ago are removed.}
}
\value{
Invisibly returns the number of files deleted.
}
\description{
Deletes locally cached data files downloaded by the hmrc package.
By default, all cached files are removed. Use \code{max_age_days} to
remove only files older than a given number of days.
}
\examples{
\donttest{
op <- options(hmrc.cache_dir = tempdir())
# Remove all cached files
clear_cache()

# Remove files older than 30 days
clear_cache(max_age_days = 30)
options(op)
}

}
\seealso{
Other data access: 
\code{\link{get_tax_gap}()},
\code{\link{list_tax_heads}()}
}
\concept{data access}
