% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec.R
\name{deprecated}
\alias{deprecated}
\alias{get_tax_receipts}
\alias{get_vat}
\alias{get_fuel_duties}
\alias{get_tobacco_duties}
\alias{get_corporation_tax}
\alias{get_stamp_duty}
\alias{get_property_transactions}
\alias{get_rd_credits}
\alias{get_tax_gap}
\alias{get_income_tax_stats}
\alias{list_tax_heads}
\alias{clear_cache}
\title{Deprecated get_* aliases}
\usage{
get_tax_receipts(...)

get_vat(...)

get_fuel_duties(...)

get_tobacco_duties(...)

get_corporation_tax(...)

get_stamp_duty(...)

get_property_transactions(...)

get_rd_credits(...)

get_tax_gap(...)

get_income_tax_stats(...)

list_tax_heads()

clear_cache(max_age_days = NULL)
}
\description{
Renamed to use the \verb{hmrc_*} prefix in v0.4.0. These aliases will be
removed in v0.6.0.
}
\keyword{internal}
